--- src/types/midgard.ts
@@ -1,11 +1,0 @@
-/**
- * This modules includes types for Midgard,
- * which are not part of any endpoint
- * and therefore not generated by `openapi-generator`
- */
-
-export type Asset = {
-  chain?: string
-  symbol?: string
-  ticker?: string
-}
